requirements.txt
fastapi==0.111.0
langchain==0.2.6
langchain_anthropic==0.1.19
langchain_community==0.2.6
langchain_core==0.2.11
langchain_openai==0.1.14
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.1
